Project reorganization, updating view state, and adding rec provider
